+++ crates/macros/src/schematic/mod.rs
@@ -1,20 +1,29 @@
-// use crate::common_schema::*;
-// use darling::{FromDeriveInput, FromMeta};
-// use proc_macro2::{Ident, TokenStream};
-// use quote::{format_ident, quote, ToTokens};
-// use syn::{Attribute, ExprPath};
+use crate::common::Macro;
+use crate::utils::extract_comment;
+use proc_macro2::TokenStream;
+use quote::{quote, ToTokens};
 
-// // #[schematic()]
-// #[derive(FromDeriveInput, Default)]
-// #[darling(default, attributes(schematic), supports(struct_named, enum_any))]
-// pub struct SchematicArgs {
-//     allow_unknown_fields: bool,
-//     context: Option<ExprPath>,
-//     env_prefix: Option<String>,
-//     file: Option<String>,
+pub struct SchematicMacro<'l>(pub Macro<'l>);
 
-//     // serde
-//     rename: Option<String>,
-//     rename_all: Option<String>,
-//     serde: SerdeMeta,
-// }
+impl<'l> ToTokens for SchematicMacro<'l> {
+    fn to_tokens(&self, tokens: &mut TokenStream) {
+        let cfg = &self.0;
+        let name = cfg.name;
+        let schema = cfg.type_of.generate_schema(
+            name,
+            extract_comment(&cfg.attrs),
+            cfg.get_casing_format(),
+            cfg.get_tagged_format(),
+        );
+
+        tokens.extend(quote! {
+            #[automatically_derived]
+            impl schematic::Schematic for #name {
+                fn generate_schema() -> schematic::SchemaType {
+                    use schematic::schema::*;
+                    #schema
+                }
+            }
+        });
+    }
+}