+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+#### 🚀 Updates
+
+- Added a `Schematic` derive macro that only implements the `schematic::Schematic` trait.
+
 #### ⚙️ Internal
 
 - Updated Rust to v1.73.